- jbj: logio: hack up a #define for bdb built w/i --with-uniquename=_rpmdb. - jbj: rpmsql: test version to avoid sqlite3_sourcid() AWOL symbol. - jbj: rpmsql: hack-o-round -lsqlite3 disabled modules loading (CM14).

+++ rpmio/rpmio.c
@@ -1874,7 +1874,7 @@ if (_ftp_debug)
 fprintf(stderr, "-> %s", req);
 
     len = strlen(req);
-    if (fdWrite(ctrl, req, len) != len) {
+    if (fdWrite(ctrl, req, len) != (int)len) {
 	rc = FTPERR_SERVER_IO_ERROR;
 	goto errxit;
     }
